+++ src/template/components/provider/src/index.js
@@ -12,7 +12,7 @@
 const provider = {
   type: 'provider',
   name: 'koop-cli-new-provider',
-  version: '1.0.0',
+  version: '0.1.0',
   hosts: false, // if true, also adds disableIdParam
   disableIdParam: true, // if true, adds to path and req.params
   Model: require('./model')